fix prospector errors
