+++ backend/battle/battles/base_stats.py
@@ -1,5 +1,5 @@
 from pokemon.helpers import get_pokemon_from_api
-
+from battle.models import Team
 
 def get_total_point_pokemon(data_pokemons):
     pokemon_first = get_pokemon_from_api(data_pokemons[0])
@@ -33,3 +33,18 @@ def sum_all_pokemons(pokemons):
 def sumValid(pokemon):
     sum_result = pokemon["attack"] + pokemon["defense"] + pokemon["hp"]
     return sum_result
+
+
+def verify_team_exists(battle, trainer):
+    pokemons_team = {
+        "pokemon_1": 0,
+        "pokemon_2": 0,
+        "pokemon_3": 0,
+    }
+    team = Team.objects.filter(battle=battle, trainer=trainer)
+    if team:
+        pokemons_team_query = team[0].pokemons.all()
+        pokemons_team["pokemon_1"] = pokemons_team_query[0]
+        pokemons_team["pokemon_2"] = pokemons_team_query[1]
+        pokemons_team["pokemon_3"] = pokemons_team_query[2]
+    return pokemons_team
