+++ backend/battle/views.py
@@ -7,7 +7,7 @@ from django.shortcuts import get_object_or_404
 from battle.models import Battle, Team
 from battle.forms import TeamForm, BattleForm, UserForm
 from battle.battles.battle import run_battle, set_winner
-
+from battle.battles.base_stats import verify_team_exists
 from users.models import User
 
 
@@ -75,24 +75,11 @@ class BattleDetail(DetailView):
         team_user = Team.objects.filter(battle=self.object, trainer=self.request.user)
         team_creator = verify_team_exists(self.object, self.object.creator)
         team_opponent = verify_team_exists(self.object, self.object.opponent)
-        context['team_creator'] = team_creator            
+        context['team_creator'] = team_creator   
         context['team_opponent'] = team_opponent
         context['team_user'] = team_user
         return context
 
-def verify_team_exists(battle, trainer):
-    pokemons_team = {
-        "pokemon_1":0,
-        "pokemon_2":0,
-        "pokemon_3":0,
-    }
-    team = Team.objects.filter(battle=battle, trainer=trainer)
-    if team:
-        pokemons_team_query = team[0].pokemons.all()
-        pokemons_team["pokemon_1"] = pokemons_team_query[0]
-        pokemons_team["pokemon_2"] = pokemons_team_query[1]
-        pokemons_team["pokemon_3"] = pokemons_team_query[2]
-    return pokemons_team
 
 class BattleSignUp(CreateView):
     model = User
